+++ R/plotDEAnalysis.R
@@ -290,6 +290,14 @@ plotDEGRegression <- function(inSCE, useResult, threshP = FALSE, labelBy = NULL,
 #' log2FC larger than this value. Default \code{0.25}.
 #' @param fdrThreshold numeric. Only fetch DEGs with FDR value smaller than this
 #' value. Default \code{0.05}.
+#' @param minGroup1MeanExp numeric. Only fetch DEGs with mean expression in 
+#' group1 greater then this value. Default \code{NULL}.
+#' @param maxGroup2MeanExp numeric. Only fetch DEGs with mean expression in 
+#' group2 less then this value. Default \code{NULL}.
+#' @param minGroup1ExprPerc numeric. Only fetch DEGs expressed in greater then 
+#' this fraction of cells in group1. Default \code{NULL}.
+#' @param maxGroup2ExprPerc numeric. Only fetch DEGs expressed in less then this 
+#' fraction of cells in group2. Default \code{NULL}.
 #' @return A \code{data.frame} object of the top DEGs, with variables of 
 #' \code{Gene}, \code{Log2_FC}, \code{Pvalue}, and \code{FDR}.
 #' @export
@@ -302,7 +310,9 @@ plotDEGRegression <- function(inSCE, useResult, threshP = FALSE, labelBy = NULL,
 #'                    analysisName = "w.aVSb")
 #' getDEGTopTable(sce.w, "w.aVSb")
 getDEGTopTable <- function(inSCE, useResult, labelBy = NULL, onlyPos = FALSE,
-                        log2fcThreshold = 0.25, fdrThreshold = 0.05){
+                        log2fcThreshold = 0.25, fdrThreshold = 0.05,
+                        minGroup1MeanExp = NULL, maxGroup2MeanExp = NULL, 
+                        minGroup1ExprPerc = NULL, maxGroup2ExprPerc = NULL){
   # Check
   .checkDiffExpResultExists(inSCE, useResult, labelBy)
   # Extract
@@ -312,15 +322,9 @@ getDEGTopTable <- function(inSCE, useResult, labelBy = NULL, onlyPos = FALSE,
     result$Gene <- SummarizedExperiment::rowData(inSCE[genes,])[[labelBy]]
   }
   # Filter
-  if (isTRUE(onlyPos)) {
-    result <- result[result$Log2_FC > 0,]
-  }
-  if (!is.null(log2fcThreshold)) {
-    result <- result[abs(result$Log2_FC) > log2fcThreshold,]
-  }
-  if (!is.null(fdrThreshold)) {
-    result <- result[result$FDR < fdrThreshold,]
-  }
+  result <- .filterDETable(result, onlyPos, log2fcThreshold, fdrThreshold, 
+                           minGroup1MeanExp, maxGroup2MeanExp, 
+                           minGroup1ExprPerc, maxGroup2ExprPerc)
   return(result)
 }
 
@@ -345,6 +349,14 @@ getDEGTopTable <- function(inSCE, useResult, labelBy = NULL, onlyPos = FALSE,
 #' log2FC larger than this value. Default \code{0.25}.
 #' @param fdrThreshold numeric. Only plot DEGs with FDR value smaller than this
 #' value. Default \code{0.05}.
+#' @param minGroup1MeanExp numeric. Only plot DEGs with mean expression in 
+#' group1 greater then this value. Default \code{NULL}.
+#' @param maxGroup2MeanExp numeric. Only plot DEGs with mean expression in 
+#' group2 less then this value. Default \code{NULL}.
+#' @param minGroup1ExprPerc numeric. Only plot DEGs expressed in greater then 
+#' this fraction of cells in group1. Default \code{NULL}.
+#' @param maxGroup2ExprPerc numeric. Only plot DEGs expressed in less then this 
+#' fraction of cells in group2. Default \code{NULL}.
 #' @param useAssay character. A string specifying an assay of expression value
 #' to plot. By default the assay used for \code{runMAST()} will be used.
 #' Default \code{NULL}.
@@ -383,11 +395,13 @@ getDEGTopTable <- function(inSCE, useResult, labelBy = NULL, onlyPos = FALSE,
 #'                    groupName1 = "w.alpha", groupName2 = "w.beta",
 #'                    analysisName = "w.aVSb")
 #' plotDEGHeatmap(sce.w, "w.aVSb")
-#' @return A \code{ComplexHeatmap::Heatmap} object
+#' @return A \code{\link[ggplot2]{ggplot}} object
 #' @export
 #' @author Yichen Wang
 plotDEGHeatmap <- function(inSCE, useResult, doLog = FALSE, onlyPos = FALSE,
                            log2fcThreshold = 0.25, fdrThreshold = 0.05,
+                           minGroup1MeanExp = NULL, maxGroup2MeanExp = NULL, 
+                           minGroup1ExprPerc = NULL, maxGroup2ExprPerc = NULL,
                            useAssay = NULL, featureAnnotations = NULL,
                            cellAnnotations = NULL,
                            featureAnnotationColor = NULL,
@@ -426,7 +440,9 @@ plotDEGHeatmap <- function(inSCE, useResult, doLog = FALSE, onlyPos = FALSE,
   deg.filtered <- getDEGTopTable(inSCE, useResult = useResult, labelBy = NULL, 
                                  onlyPos = onlyPos, 
                                  log2fcThreshold = log2fcThreshold, 
-                                 fdrThreshold = fdrThreshold)
+                                 fdrThreshold = fdrThreshold,
+                                 minGroup1MeanExp, maxGroup2MeanExp, 
+                                 minGroup1ExprPerc, maxGroup2ExprPerc)
   if(dim(deg.filtered)[1] <= 1){
     stop('Too few genes that pass filtration, unable to plot')
   }